habit-tracker/pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "flask>=3.1.3",
    "pymongo[srv]>=4.17.0",
    "python-dotenv>=1.2.2",
]
